Open CRAN PDF manual of a package

PDF versions of the manual are usually not included as vignettes in R packages. Still this format is convenient for reading and doing full text search.
This function creates the appropriate link to the pdf file on CRAN and opens the pdf manual in a browser window.

Usage

pdfManual(package)

Arguments

package

Package name (symbol or character)

See also